2021/8: optimize part 2 solution by not trying impossible mappings
The previous algorithm tries literally every possible arrangement of
a-f, ignoring that repetitions of characters cannot be valid: no segment
is in two or more places.

Now, permutations of a-g are generated, which use each segment exactly once.

This reduces the search space by a factor of ~416: from 8^8 (16777216)
to 8! (40320), and also the runtime by a factor of ~100: from ~40 seconds
to ~400 milliseconds.