2015, 2016, 2022, lib: rewrite to not use io.Reader in the Solver
io.Reader is a vestigal artifact from when the input file would be
directly passed to the solver to read the input.

Nowadays the file is also read from the API (all the way, for
simplicity) so we always have the whole thing, and the reader
abstraction isn't needed.

This allows for parse functions to be re-used inside parse functions for
certain days, which can be useful when parsing unique chunks.